Criado processo de comunicação com o firestore
