+++ src/types/User/PerfilType.ts
@@ -4,3 +4,10 @@ export type PerfilType = {
     user: UserType,
     setViewPerfil: (viewPerfil: boolean) => void
 }
+
+export type PerfilProps = {
+    show: boolean, 
+    setShow: (show: boolean) => void,
+    user: UserType,
+    setUser: (user: UserType) => void
+}
